% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esp-check-access.R
\encoding{UTF-8}
\name{esp_check_access}
\alias{esp_check_access}
\title{Check access to SIANE data resources}
\usage{
esp_check_access()
}
\value{
Logical scalar, \code{TRUE} if accessible and \code{FALSE} otherwise.
}
\description{
Check whether \strong{R} has access to resources at
\url{https://github.com/rOpenSpain/mapSpain/tree/sianedata}.
}
\examples{

esp_check_access()
}
\seealso{
\code{\link[giscoR:gisco_check_access]{giscoR::gisco_check_access()}}.
}
\keyword{internal}
