% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esp-get-attributions.R
\encoding{UTF-8}
\name{esp_get_attributions}
\alias{esp_get_attributions}
\title{Get the attribution for a tile provider}
\usage{
esp_get_attributions(type, options = NULL)
}
\arguments{
\item{type}{This argument can be one of:
\itemize{
\item The name of one of the pre-defined providers (see
\link{esp_tiles_providers}).
\item A list with two named elements \code{id} and \code{q} with your own arguments. See
\code{\link[=esp_make_provider]{esp_make_provider()}} and examples.
}}

\item{options}{A named list containing additional options to pass to the
query.}
}
\value{
A character string with the provider attribution, or \code{NULL} if no
attribution is available.
}
\description{
\code{esp_get_attributions()} gets the attribution of a tile provider defined by
the \code{type} argument.
}
\examples{
esp_get_attributions("IGNBase.Todo")
}
\seealso{
\itemize{
\item \code{\link[=esp_get_tiles]{esp_get_tiles()}} downloads static map tiles.
\item \code{\link[giscoR:gisco_attributions]{giscoR::gisco_attributions()}} provides GISCO attribution text.
}

Static map tiles and imagery:
\code{\link[=addProviderEspTiles]{addProviderEspTiles()}},
\code{\link[=esp_get_tiles]{esp_get_tiles()}},
\code{\link[=esp_make_provider]{esp_make_provider()}}
}
\concept{images}
