% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esp-get-grid-ESDAC.R
\encoding{UTF-8}
\name{esp_get_grid_ESDAC}
\alias{esp_get_grid_ESDAC}
\title{National geographic grids from the European Soil Data Centre (ESDAC)}
\source{
\href{https://esdac.jrc.ec.europa.eu/content/european-reference-grids}{EEA reference grid}.
}
\usage{
esp_get_grid_ESDAC(
  resolution = c(10, 1),
  update_cache = FALSE,
  cache_dir = NULL,
  verbose = FALSE
)
}
\arguments{
\item{resolution}{Numeric. Resolution of the grid in kilometers. Can be \code{1}
or \code{10}.}

\item{update_cache}{Logical. If \code{TRUE}, refreshes the cached file and forces
a new download. Defaults to \code{FALSE}.}

\item{cache_dir}{Character string. A path to a cache directory. See
\strong{Caching}.}

\item{verbose}{A logical value. If \code{TRUE} displays informational messages.}
}
\value{
An \code{\link[sf:st_sf]{sf}} \code{POLYGON} object.
}
\description{
This function loads an \code{\link[sf:st_sf]{sf}} \code{POLYGON} with the geographic
grids of Spain provided by the European Soil Data Centre (ESDAC).
}
\section{Caching}{


Functions that download data store files in \code{cache_dir}. When \code{cache_dir}
is \code{NULL}, they use the active package cache, which defaults to a temporary
directory. Set \code{update_cache = TRUE} to replace an existing cached file.
See \strong{Caching strategies} in \code{\link[=esp_set_cache_dir]{esp_set_cache_dir()}} to configure a
persistent cache.
}

\examples{
\dontshow{if (esp_check_access()) withAutoprint(\{ # examplesIf}
\dontrun{
grid <- esp_get_grid_ESDAC()
esp <- esp_get_spain(moveCAN = FALSE)

library(ggplot2)

ggplot(grid) +
  geom_sf() +
  geom_sf(data = esp, color = "grey50", fill = NA) +
  theme_light() +
  labs(title = "ESDAC Grid for Spain")
}
\dontshow{\}) # examplesIf}
}
\references{
\itemize{
\item Panagos P., Van Liedekerke M., Jones A., Montanarella L., "European Soil
Data Centre: Response to European policy support and public data
requirements", (2012) \emph{Land Use Policy}, 29 (2), pp. 329-338.
\doi{10.1016/j.landusepol.2011.07.003}.
\item European Soil Data Centre (ESDAC), esdac.jrc.ec.europa.eu, European
Commission, Joint Research Centre.
}
}
\seealso{
Geographic grid data:
\code{\link[=esp_get_grid_BDN]{esp_get_grid_BDN()}},
\code{\link[=esp_get_grid_MTN]{esp_get_grid_MTN()}}
}
\concept{grids}
