% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esp-get-gridmap.R
\encoding{UTF-8}
\name{esp_get_gridmap}
\alias{esp_get_gridmap}
\alias{esp_get_hex_prov}
\alias{esp_get_hex_ccaa}
\alias{esp_get_grid_prov}
\alias{esp_get_grid_ccaa}
\title{Hexbin or square grid maps of Spain}
\usage{
esp_get_hex_prov(prov = NULL)

esp_get_hex_ccaa(ccaa = NULL)

esp_get_grid_prov(prov = NULL)

esp_get_grid_ccaa(ccaa = NULL)
}
\arguments{
\item{prov, ccaa}{Character. A vector of names, codes or both for provinces
and Autonomous Communities and Cities, or \code{NULL} to get all the data. See
\strong{Details}.}
}
\value{
An \code{\link[sf:st_sf]{sf}} object.
}
\description{
This function loads a hexbin map (\code{\link[sf:st_sf]{sf}} object) or a map of
squares with boundaries of the provinces or Autonomous Communities and
Cities of Spain.
}
\details{
Hexbin (or grid) maps have an advantage over traditional choropleth maps.
In choropleths, regions with larger polygons tend to appear more prominent
simply because of their size, which introduces visual bias. With hexbin
maps, each region is represented equally, reducing this bias.

You can use and mix names, ISO codes, \code{"codauto"} or \code{"cpro"} codes (see
\link{esp_codelist}) and NUTS codes of different levels.

When using a code corresponding to a higher level (for example,
\code{esp_get_prov("Andalucia")}), all the corresponding units of that level are
provided (in this case, all the provinces of Andalusia).

Results are provided in \strong{EPSG:4258}. Use \code{\link[sf:st_transform]{sf::st_transform()}}
to change the projection.
}
\examples{
\dontshow{if (esp_check_access()) withAutoprint(\{ # examplesIf}
\donttest{
esp <- esp_get_spain()
hexccaa <- esp_get_hex_ccaa()

library(ggplot2)

ggplot(hexccaa) +
  geom_sf(data = esp) +
  geom_sf(aes(fill = codauto), alpha = 0.3, show.legend = FALSE) +
  geom_sf_text(aes(label = label), check_overlap = TRUE) +
  theme_void() +
  labs(title = "Hexbin: Autonomous Communities and Cities")

hexprov <- esp_get_hex_prov()

ggplot(hexprov) +
  geom_sf(data = esp) +
  geom_sf(aes(fill = codauto), alpha = 0.3, show.legend = FALSE) +
  geom_sf_text(aes(label = label), check_overlap = TRUE) +
  theme_void() +
  labs(title = "Hexbin: Provinces")

gridccaa <- esp_get_grid_ccaa()

ggplot(gridccaa) +
  geom_sf(data = esp) +
  geom_sf(aes(fill = codauto), alpha = 0.3, show.legend = FALSE) +
  geom_sf_text(aes(label = label), check_overlap = TRUE) +
  theme_void() +
  labs(title = "Grid: Autonomous Communities and Cities")

gridprov <- esp_get_grid_prov()

ggplot(gridprov) +
  geom_sf(data = esp) +
  geom_sf(aes(fill = codauto), alpha = 0.3, show.legend = FALSE) +
  geom_sf_text(aes(label = label), check_overlap = TRUE) +
  theme_void() +
  labs(title = "Grid: Provinces")
}
\dontshow{\}) # examplesIf}
}
\seealso{
Additional boundary datasets and representations:
\code{\link[=esp_get_comarca]{esp_get_comarca()}},
\code{\link{esp_get_simpl}}
}
\concept{boundary_helpers}
\concept{political}
