% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esp_get_hex.R
\name{esp_get_hex}
\alias{esp_get_hex}
\alias{esp_get_hex_prov}
\alias{esp_get_hex_ccaa}
\title{Get an hexbin map with the boundaries of Spain}
\usage{
esp_get_hex_prov(prov = NULL)

esp_get_hex_ccaa(ccaa = NULL)
}
\arguments{
\item{prov}{See \link{esp_get_prov}}

\item{ccaa}{See \link{esp_get_ccaa}}
}
\value{
A \code{POLYGON} object.
}
\description{
Loads a hexbin map (\code{sf} object) with the
boundaries of the provinces or autonomous communities of Spain.
}
\details{
Hexbin or grid map has an advantage over usual choropleth maps.
In choropleths, a large polygon data looks more emphasized just because
of its size, what introduces a bias. Here with hexbin, each region is
represented equally dismissing the bias.
}
\examples{

library(sf)


ccaa <- esp_get_hex_ccaa()
plot(st_geometry(ccaa), col = hcl.colors(4))

prov <- esp_get_hex_prov()
plot(st_geometry(prov), col = hcl.colors(15))
}
\seealso{
\link{esp_get_nuts}, \link{esp_get_ccaa}, \link{esp_get_prov},
\link{esp_get_munic}, \link{esp_codelist}
}
\author{
dieghernan, \url{https://github.com/dieghernan/}
}
