% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esp_get_rivers.R
\name{esp_get_rivers}
\alias{esp_get_rivers}
\title{Get rivers, channels, reservoirs and other wetlands of Spain}
\source{
IGN data via a custom CDN (see
\url{https://github.com/rOpenSpain/mapSpain/tree/sianedata}).
}
\usage{
esp_get_rivers(
  epsg = "4258",
  cache = TRUE,
  update_cache = FALSE,
  cache_dir = NULL,
  verbose = FALSE,
  resolution = 3,
  spatialtype = "line",
  name = NULL
)
}
\arguments{
\item{epsg, cache, update_cache, cache_dir, verbose}{See
\code{\link{esp_get_nuts}}.}

\item{resolution}{Resolution of the polygon. Values available are
\code{"3", "6.5"} or  \code{"10"}.}

\item{spatialtype}{Spatial type of the output. Use \code{"area"} for
\code{POLYGONS} or \code{"line"} for \code{LINESTRING}.}

\item{name}{Optional. A character or regex expresion with the name of the
element to be extracted. See Details}
}
\value{
A \code{POLYGON} of \code{LINESTRING} object.
}
\description{
Loads a simple feature (\code{sf}) object containing lines or
areas with the required hydrograpic elements of Spain.
}
\details{
Metadata available on \url{https://github.com/rOpenSpain/mapSpain/tree/sianedata/data-raw/documentacion_cartosiane}.

\code{name} admits regex expresions. See \code{\link[base]{regex}} for more
information.
}
\examples{
library(sf)

# Use of regex

regex1 <- esp_get_rivers(name = "Tajo|Segura")
unique(regex1$rotulo)

regex2 <- esp_get_rivers(name = "Tajo$| Segura")
unique(regex2$rotulo)

# Rivers in Spain
shapeEsp <- esp_get_country(moveCAN = FALSE)

MainRivers <-
  esp_get_rivers(name = "Tajo$|Ebro$|Ebre$|Duero|Guadiana$|Guadalquivir")

opar <- par(no.readonly = TRUE)
par(mar = c(0, 0, 0, 0))

plot(st_geometry(MainRivers), col = "skyblue", lwd = 1.5)
plot(st_geometry(shapeEsp), col = NA, add = TRUE)

# All wetlands

Wetlands <- esp_get_rivers(spatialtype = "area")
plot(st_geometry(Wetlands), col = "skyblue", border = NA)
plot(st_geometry(shapeEsp), col = NA, add = TRUE)

par(opar)
}
\author{
dieghernan, \url{https://github.com/dieghernan/}
}
\concept{mapnatural}
