% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\encoding{UTF-8}
\name{esp_munic.sf}
\alias{esp_munic.sf}
\title{All Municipalities \code{POLYGON} object of Spain}
\format{
A \code{POLYGON} data frame (resolution: 1:1million, EPSG:4258) object:
\itemize{
\item codauto: INE code of each autonomous community.
\item ine.ccaa.name: INE name of each autonomous community.
\item cpro: INE code of each province.
\item ine.prov.name: INE name of each province.
\item cmun: INE code of each municipality.
\item name: Name of the municipality.
\item LAU_CODE: LAU Code (GISCO) of the municipality.
\item geometry: geometry field.
}
}
\source{
\url{https://ec.europa.eu/eurostat/web/gisco/geodata/reference-data/}, LAU 2019
data.
}
\description{
A \code{sf} object including all municipalities of Spain as provided by GISCO
(2019 version).
}
\examples{
data("esp_munic.sf")
data("esp_nuts.sf")

Teruel.cpro <- esp_dict_region_code("Teruel", destination = "cpro")
Teruel.NUTS <- esp_dict_region_code(Teruel.cpro,
  origin = "cpro",
  destination = "nuts"
)

Teruel.sf <- esp_munic.sf[esp_munic.sf$cpro == Teruel.cpro, ]
Teruel.city <- Teruel.sf[Teruel.sf$name == "Teruel", ]

# Extract CCAA

CCAA <- esp_get_prov()


library(tmap)

tm_shape(CCAA, bbox = Teruel.sf) +
  tm_polygons("wheat") +
  tm_shape(Teruel.sf) +
  tm_polygons("cornsilk") +
  tm_shape(Teruel.city) +
  tm_fill("firebrick") +
  tm_layout(main.title = "Municipalities of Teruel")
}
\seealso{
\code{\link[=esp_get_munic]{esp_get_munic()}}
}
\concept{datasets}
