% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esp-check-access.R
\name{esp_check_access}
\alias{esp_check_access}
\title{Check access to SIANE data}
\usage{
esp_check_access()
}
\value{
A logical.
}
\description{
Check if \strong{R} has access to resources at
\url{https://github.com/rOpenSpain/mapSpain/tree/sianedata}.
}
\examples{

esp_check_access()
}
\seealso{
\code{\link[giscoR:gisco_check_access]{giscoR::gisco_check_access()}}.
}
\keyword{internal}
