% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\encoding{UTF-8}
\name{esp_codelist}
\alias{esp_codelist}
\title{Spanish Code Translation Data Frame}
\format{
data frame with codes as columns
\itemize{
\item \strong{nuts*.code}: NUTS code of each subdivision.
\item \strong{nuts*.code}: NUTS name of each subdivision.
\item \strong{codauto}: INE code of each autonomous community.
\item \strong{iso2.*.code}: ISO2 code of each autonomous
community and province.
\item \strong{ine.*.name}: INE name of each autonomous community
and province.
\item \strong{iso2.*.name.(lang)}: ISO2 name of each autonomous community
and province. Several languages available.
\item \strong{cldr.*.name.(lang)}: CLDR name of each autonomous community and
province. Several languages available.
\item \strong{ccaa.short.*}: Short (common) name of each autonomous
community. Several languages available.
\item \strong{cpro}: INE code of each province.
\item \strong{prov.shortname.*}: Short (common) name of each province.
Several languages available.
}
}
\source{
\itemize{
\item INE: Instituto Nacional de Estadistica: \url{https://www.ine.es/}
\item Eurostat (NUTS): \url{https://ec.europa.eu/eurostat/web/nuts/background}
\item ISO: \url{https://www.iso.org/obp/ui/#iso:code:3166:ES}
\item CLDR: \url{https://unicode-org.github.io/cldr-staging/charts/38/index.html}
}
}
\description{
A data frame used internally for translating codes and names of the
different subdivisions of Spain. The data frame provides the hierarchy of
the subdivisions including NUTS1 level, Autonomous Communities
(equivalent to NUTS2), Provinces and NUTS3 level. See Note.
}
\note{
Languages available are:
\itemize{
\item "en": English
\item "es": Spanish
\item "ca": Catalan
\item "ga": Galician
\item "eu": Basque
}

Although NUTS2 matches the first subdivision level of Spain
(CCAA - Autonomous Communities), it should be noted that NUTS3 does not
match the second subdivision level of Spain (Provinces). NUTS3 provides a
dedicated code for major islands whereas the Provinces doesn't.

Ceuta and Melilla has an specific status (Autonomous Cities) but are
considered as communities with a single province (as Madrid, Asturias
or Murcia) on this dataset.
}
\examples{

library(tibble)

glimpse(as_tibble(esp_codelist))
}
\concept{datasets}
