% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esp_get_capimun.R
\name{esp_get_capimun}
\alias{esp_get_capimun}
\title{Get the location of municipalities of Spain}
\source{
IGN data via a custom CDN (see
\url{https://github.com/rOpenSpain/mapSpain/tree/sianedata}.
}
\usage{
esp_get_capimun(
  year = Sys.Date(),
  epsg = "4258",
  cache = TRUE,
  update_cache = FALSE,
  cache_dir = NULL,
  verbose = FALSE,
  region = NULL,
  munic = NULL,
  moveCAN = TRUE,
  rawcols = FALSE
)
}
\arguments{
\item{year}{Release year. See Details for years available.}

\item{epsg}{projection of the map: 4-digit
\href{https://epsg.io/}{EPSG code}. One of:
\itemize{
\item \code{"4258"} - \href{https://epsg.io/4258}{ETRS89}
\item \code{"4326"} - \href{https://epsg.io/4326}{WGS84}
\item \code{"3035"} - \href{https://epsg.io/3035}{ETRS89 / ETRS-LAEA}
\item \code{"3857"} - \href{https://epsg.io/3857}{Pseudo-Mercator}
}}

\item{cache}{A logical whether to do caching. Default is \code{TRUE}.}

\item{update_cache}{A logical whether to update cache.
Default is \code{FALSE}. When set to \code{TRUE} it would force a fresh
download of the source \code{.geojson} file.}

\item{cache_dir}{A path to a cache directory. See Details.}

\item{verbose}{Display information. Useful for debugging,
default is \code{FALSE}.}

\item{region}{Optional. A vector of region names, NUTS or ISO codes
(see \link{esp_dict_region_code}).}

\item{munic}{A name or regex expression with the names of the required
municipalities. \code{NULL} would not produce any filtering.}

\item{moveCAN}{A logical \code{TRUE,FALSE} or a vector of coordinates
\code{c(lat, lon)}. It places the Canary Island close to Spain's mainland.
Initial position can be adjusted using the vector of coordinates.
See Note.}

\item{rawcols}{Logical. Setting this to \code{TRUE} would add the raw
columns of the dataset provided by IGN.}
}
\value{
A \code{POINT} object.
}
\description{
Get the location of the political powers for each municipality (possibly
the center of the municipality).

Note that this differs of the centroid of the boundaries of the
municipallity, returned by \code{\link[=esp_get_munic]{esp_get_munic()}}.
}
\details{
\code{year} could be passed as a single year ("YYYY" format, as end of year) or
as a specific date ("YYYY-MM-DD" format).
Historical information starts as of 2005.

When using \code{region} you can use and mix names and NUTS codes (levels 1,
2 or 3), ISO codes (corresponding to level 2 or 3) or \code{cpro}.

When calling a superior level (Province, Autonomous Community or NUTS1) ,
all the municipalities of that level would be added.
}
\seealso{
\code{\link[=esp_get_munic]{esp_get_munic()}}, \link{esp_munic.sf}, \link{esp_codelist}
}
\author{
dieghernan, \url{https://github.com/dieghernan/}.
}
\concept{mappolitical}
