% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esp-get-ccaa.R
\encoding{UTF-8}
\name{esp_get_ccaa}
\alias{esp_get_ccaa}
\title{Autonomous Communities and Cities of Spain from GISCO}
\source{
\url{https://gisco-services.ec.europa.eu/distribution/v2/}.

Copyright:
\url{https://ec.europa.eu/eurostat/web/gisco/geodata/administrative-units}.
}
\usage{
esp_get_ccaa(ccaa = NULL, moveCAN = TRUE, ...)
}
\arguments{
\item{ccaa}{Character string. A vector of names, codes or both for
Autonomous Communities and Cities, or \code{NULL} to get all Autonomous
Communities and Cities.
See \strong{Details}.}

\item{moveCAN}{A logical \code{TRUE/FALSE} or a vector of coordinates
\code{c(lat, lon)}. It places the Canary Islands close to Spain's mainland.
Initial position can be adjusted using the vector of coordinates. See
\strong{Displacing the Canary Islands} in \code{\link[=esp_move_can]{esp_move_can()}}.}

\item{...}{
  Arguments passed on to \code{\link{esp_get_nuts}}
  \describe{
    \item{\code{year}}{Year character string or number. Release year of the file. See
\code{\link[giscoR:gisco_get_nuts]{giscoR::gisco_get_nuts()}} for valid values.}
    \item{\code{epsg}}{Character string or number. Projection of the map: 4-digit
\href{https://epsg.io/}{EPSG code}. One of:
\itemize{
\item \code{"4258"}: \href{https://epsg.io/4258}{ETRS89}.
\item \code{"4326"}: \href{https://epsg.io/4326}{WGS84}.
\item \code{"3035"}: \href{https://epsg.io/3035}{ETRS89 / ETRS-LAEA}.
\item \code{"3857"}: \href{https://epsg.io/3857}{Pseudo-Mercator}.
}}
    \item{\code{cache}}{Logical. Whether to cache downloaded files. Defaults to \code{TRUE}.
See \strong{Caching}.}
    \item{\code{update_cache}}{Logical. If \code{TRUE}, refreshes the cached file and forces
a new download. Defaults to \code{FALSE}.}
    \item{\code{cache_dir}}{Character string. A path to a cache directory. See
\strong{Caching}.}
    \item{\code{spatialtype}}{Character string. Type of geometry to be returned.
Options available are:
\itemize{
\item \code{"RG"}: regions, returned as a \code{MULTIPOLYGON/POLYGON} object.
\item \code{"LB"}: labels, returned as a \code{POINT} object.
}}
    \item{\code{ext}}{Character. Extension of the file (default \code{"gpkg"}). See
\code{\link[giscoR:gisco_get_nuts]{giscoR::gisco_get_nuts()}}.}
    \item{\code{verbose}}{A logical value. If \code{TRUE} displays informational messages.}
    \item{\code{resolution}}{A character string or numeric value with the geospatial
data resolution. One of:
\itemize{
\item \code{"60"}: 1:60 million.
\item \code{"20"}: 1:20 million.
\item \code{"10"}: 1:10 million.
\item \code{"03"}: 1:3 million.
\item \code{"01"}: 1:1 million.
}}
  }}
}
\value{
A \code{\link[sf:st_sf]{sf}} object.
}
\description{
Get
\href{https://en.wikipedia.org/wiki/Autonomous_communities_of_Spain}{Autonomous Communities and Cities of Spain} at a
specified scale.
}
\details{
When using \code{ccaa} you can use and mix names and NUTS codes (levels 1 or 2),
ISO codes (corresponding to level 2) or \code{codauto} (see \link{esp_codelist}).
Ceuta and Melilla are included at the Autonomous Communities and Cities
level in this function.

When calling a NUTS 1 level, all Autonomous Communities and Cities of that
level are added.
}
\section{Caching}{


Functions that download data store files in \code{cache_dir}. When \code{cache_dir}
is \code{NULL}, they use the active package cache, which defaults to a temporary
directory. Set \code{update_cache = TRUE} to replace an existing cached file.
See \strong{Caching strategies} in \code{\link[=esp_set_cache_dir]{esp_set_cache_dir()}} to configure a
persistent cache.
}

\examples{
ccaa <- esp_get_ccaa()

library(ggplot2)

ggplot(ccaa) +
  geom_sf()

# Random Autonomous Communities and Cities.
random_ccaa <- esp_get_ccaa(ccaa = c(
  "Euskadi",
  "Catalunya",
  "ES-EX",
  "Canarias",
  "ES52",
  "01"
))

ggplot(random_ccaa) +
  geom_sf(aes(fill = codauto), show.legend = FALSE) +
  geom_sf_label(aes(label = codauto), alpha = 0.3) +
  coord_sf(crs = 3857)

# All Autonomous Communities and Cities of a NUTS 1 region plus one.

mixed <- esp_get_ccaa(ccaa = c("La Rioja", "Noroeste"))

ggplot(mixed) +
  geom_sf()

# Combine with giscoR to get countries.
\donttest{

library(giscoR)
library(sf)

res <- 20 # Use the same resolution.

europe <- gisco_get_countries(resolution = res)
ccaa <- esp_get_ccaa(moveCAN = FALSE, resolution = res)

ggplot(europe) +
  geom_sf(fill = "#DFDFDF", color = "#656565") +
  geom_sf(data = ccaa, fill = "#FDFBEA", color = "#656565") +
  coord_sf(
    xlim = c(23, 74) * 10e4,
    ylim = c(14, 55) * 10e4,
    crs = 3035
  ) +
  theme(panel.background = element_rect(fill = "#C7E7FB"))
}
}
\seealso{
\code{\link[=esp_get_ccaa_siane]{esp_get_ccaa_siane()}}.

GISCO boundary data:
\code{\link[=esp_get_munic]{esp_get_munic()}},
\code{\link[=esp_get_nuts]{esp_get_nuts()}},
\code{\link[=esp_get_prov]{esp_get_prov()}},
\code{\link[=esp_get_spain]{esp_get_spain()}}
}
\concept{gisco}
\concept{political}
