% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esp_get_ccaa.R
\name{esp_get_ccaa}
\alias{esp_get_ccaa}
\alias{esp_get_ccaa_siane}
\title{Get Autonomous Communities boundaries of Spain}
\source{
IGN data via a custom CDN (see
\url{https://github.com/rOpenSpain/mapSpain/tree/sianedata}).
}
\usage{
esp_get_ccaa(ccaa = NULL, ...)

esp_get_ccaa_siane(
  ccaa = NULL,
  year = Sys.Date(),
  epsg = "4258",
  cache = TRUE,
  update_cache = FALSE,
  cache_dir = NULL,
  verbose = FALSE,
  resolution = "3",
  moveCAN = TRUE,
  rawcols = FALSE
)
}
\arguments{
\item{ccaa}{A vector of names and/or codes for autonomous communities
or \code{NULL} to get all the autonomous communities. See Details.}

\item{...}{Additional parameters from \link{esp_get_nuts}.}

\item{year, epsg, cache, update_cache, cache_dir, verbose, moveCAN}{See \code{\link{esp_get_nuts}}. See Details for \code{year}.}

\item{resolution}{Resolution of the polygon. Values available are
\code{"3", "6.5"} or  \code{"10"}.}

\item{rawcols}{Logical. Setting this to \code{TRUE} would add the raw
columns of the dataset provided by IGN.}
}
\value{
A \code{POLYGON/POINT} object.
}
\description{
Loads a simple feature (\code{sf}) object containing the
autonomous communities boundaries of Spain.

\code{esp_get_ccaa} uses GISCO (Eurostat) as source

\code{esp_get_ccaa_siane} use CartoBase ANE as source,
provided by Instituto Geografico Nacional (IGN),
\href{http://www.ign.es/web/ign/portal}{ign.es}. Years available are
2005 up to today.
}
\details{
When using \code{ccaa} you can use and mix names and NUTS codes
(levels 1 or 2), ISO codes (corresponding to level 2) or \code{codauto}.
Ceuta and Melilla are considered as Autonomous Communities on this dataset.

When calling a NUTS1 level, all the Autonomous Communities of that level
would be added.

On \code{esp_get_ccaa_siane}, \code{year} could be passed as a single
year ("YYYY" format, as end of year) or as a specific
date ("YYYY-MM-DD" format). Historical information starts as of 2005.
}
\examples{

library(sf)

# Random CCAA

Random <-
  esp_get_ccaa(ccaa = c("Euskadi",
                        "Catalunya",
                        "ES-EX",
                        "Canarias",
                        "ES52",
                        "01"))
plot(st_geometry(Random), col = hcl.colors(6))

# All CCAA of a Zone plus an addition

Mix <-
  esp_get_ccaa(ccaa = c("La Rioja", "Noroeste"),
               resolution = "20")
plot(
  Mix[, "nuts1.code"],
  pal = hcl.colors(2),
  key.pos = NULL,
  main = NULL,
  border = "white"
)
}
\seealso{
\link{esp_get_hex_ccaa}, \link{esp_get_nuts}, \link{esp_get_prov},
\link{esp_get_munic}, \link{esp_codelist}
}
\author{
dieghernan, \url{https://github.com/dieghernan/}
}
