% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esp-get-grid-EEA.R
\encoding{UTF-8}
\name{esp_get_grid_EEA}
\alias{esp_get_grid_EEA}
\title{National geographic grids from EEA}
\source{
\href{https://www.eea.europa.eu/en/datahub/datahubitem-view/3c362237-daa4-45e2-8c16-aaadfb1a003b}{EEA reference grid}.
}
\usage{
esp_get_grid_EEA(
  resolution = 100,
  type = "main",
  update_cache = FALSE,
  cache_dir = NULL,
  verbose = FALSE
)
}
\arguments{
\item{resolution}{Resolution of the grid in kilometers. Can be \code{1}, \code{10}
or \code{100}.}

\item{type}{Character. The geographic scope of the grid:
\itemize{
\item \code{"main"}: Mainland Spain (default).
\item \code{"canary"}: Canary Islands.
}}

\item{update_cache}{Logical. If \code{TRUE}, refreshes the cached file and forces
a new download. Defaults to \code{FALSE}.}

\item{cache_dir}{Character string. A path to a cache directory. See
\strong{Caching}.}

\item{verbose}{A logical value indicating whether to display informational
messages.}
}
\value{
An \code{\link[sf:st_sf]{sf}} \code{POLYGON} object.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#defunct}{\figure{lifecycle-defunct.svg}{options: alt='[Defunct]'}}}{\strong{[Defunct]}}

This function is defunct because the source file is no longer available.
}
\concept{deprecated functions}
\keyword{internal}
