% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esp_get_hydrobasin.R
\name{esp_get_hydrobasin}
\alias{esp_get_hydrobasin}
\title{Get \code{sf} polygons of the drainage basin demarcations of Spain}
\source{
IGN data via a custom CDN (see
\url{https://github.com/rOpenSpain/mapSpain/tree/sianedata}).
}
\usage{
esp_get_hydrobasin(
  epsg = "4258",
  cache = TRUE,
  update_cache = FALSE,
  cache_dir = NULL,
  verbose = FALSE,
  resolution = "3",
  domain = "land"
)
}
\arguments{
\item{epsg}{projection of the map: 4-digit \href{https://epsg.io/}{EPSG code}.
One of:
\itemize{
\item "4258": ETRS89
\item "4326": WGS84
\item "3035": ETRS89 / ETRS-LAEA
\item "3857": Pseudo-Mercator
}}

\item{cache}{A logical whether to do caching. Default is \code{TRUE}. See
\strong{About caching}.}

\item{update_cache}{A logical whether to update cache. Default is \code{FALSE}.
When set to \code{TRUE} it would force a fresh download of the source file.}

\item{cache_dir}{A path to a cache directory. See \strong{About caching}.}

\item{verbose}{Logical, displays information. Useful for debugging,
default is \code{FALSE}.}

\item{resolution}{Resolution of the polygon. Values available are "3", "6.5"
or "10".}

\item{domain}{Possible values are "land", that includes only
the ground part or the ground or "landsea", that includes both the ground
and the related sea waters of the basin}
}
\value{
A \code{sf} polygon object.
}
\description{
Loads a \code{sf} polygon object containing areas with the required
hydrographic elements of Spain.
}
\details{
Metadata available on
\url{https://github.com/rOpenSpain/mapSpain/tree/sianedata/}.
}
\section{About caching}{
You can set your \code{cache_dir} with \code{\link[=esp_set_cache_dir]{esp_set_cache_dir()}}.

Sometimes cached files may be corrupt. On that case, try re-downloading
the data setting \code{update_cache = TRUE}.

If you experience any problem on download, try to download the
corresponding .geojson file by any other method and save it on your
\code{cache_dir}. Use the option \code{verbose = TRUE} for debugging the API query.
}

\examples{
\donttest{
all <- esp_get_prov(moveCAN = FALSE)
hydroland <- esp_get_hydrobasin(domain = "land")
hydrolandsea <- esp_get_hydrobasin(domain = "landsea")

library(tmap)

tm_shape(hydrolandsea, bbox = c(-9.5, 35, 4.5, 44)) +
  tm_fill("skyblue4") +
  tm_shape(all) +
  tm_polygons("grey90") +
  tm_shape(hydroland) +
  tm_polygons("skyblue", alpha = 0.5, border.col = "blue") +
  tm_text(
    text = "rotulo",
    remove.overlap = TRUE,
    size = 0.5,
    fontface = "bold",
    shadow = TRUE
  ) +
  tm_layout(bg.color = "grey95")
}
}
\concept{natural}
