% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esp-get-prov-siane.R
\encoding{UTF-8}
\name{esp_get_prov_siane}
\alias{esp_get_prov_siane}
\title{Provinces of Spain from SIANE}
\source{
CartoBase ANE (Atlas Nacional de España) provided by Instituto Geográfico Nacional (IGN),
\url{http://www.ign.es/web/ign/portal}. Years available are 2005 up to today.

Copyright:
\url{https://centrodedescargas.cnig.es/CentroDescargas/cartobase-ane}

Always acknowledge authorship using the following statements:
\enumerate{
\item When the original digital product is not modified or altered, use one of
the following statements:
\itemize{
\item CartoBase ANE 2006-2024 CC-BY 4.0 ign.es.
\item CartoBase ANE 2006-2024 CC-BY 4.0 Instituto Geográfico Nacional.
}
\item When a new product is generated:
\itemize{
\item Obra derivada de CartoBase ANE 2006-2024 CC-BY 4.0 ign.es.
}
}

Data distributed through the \code{sianedata} data branch, see
\url{https://github.com/rOpenSpain/mapSpain/tree/sianedata}.
}
\usage{
esp_get_prov_siane(
  prov = NULL,
  year = Sys.Date(),
  epsg = 4258,
  cache = TRUE,
  update_cache = FALSE,
  cache_dir = NULL,
  verbose = FALSE,
  resolution = c(3, 6.5, 10),
  moveCAN = TRUE,
  rawcols = FALSE
)
}
\arguments{
\item{prov}{A vector of names, codes or both for provinces, or \code{NULL} to get
all the provinces. See \strong{Details}.}

\item{year}{Character string or number. Release year. It must use format
\code{YYYY} (assuming end of year) or \code{YYYY-MM-DD}. Historical information
starts as of 2005.}

\item{epsg}{Character string or number. Projection of the map: 4-digit
\href{https://epsg.io/}{EPSG code}. One of:
\itemize{
\item \code{"4258"}: \href{https://epsg.io/4258}{ETRS89}.
\item \code{"4326"}: \href{https://epsg.io/4326}{WGS84}.
\item \code{"3035"}: \href{https://epsg.io/3035}{ETRS89 / ETRS-LAEA}.
\item \code{"3857"}: \href{https://epsg.io/3857}{Pseudo-Mercator}.
}}

\item{cache}{Logical. Whether to cache downloaded files. Default is \code{TRUE}.
See \strong{Caching strategies} section in \code{\link[=esp_set_cache_dir]{esp_set_cache_dir()}}.}

\item{update_cache}{Logical. If \code{TRUE}, refreshes the cached file and forces
a new download. Defaults to \code{FALSE}.}

\item{cache_dir}{Character string. A path to a cache directory. See
\strong{Caching strategies} section in \code{\link[=esp_set_cache_dir]{esp_set_cache_dir()}}.}

\item{verbose}{A logical value. If \code{TRUE} displays informational messages.}

\item{resolution}{Character string or number. Resolution of the geospatial
data. One of:
\itemize{
\item \code{"10"}: 1:10 million.
\item \code{"6.5"}: 1:6.5 million.
\item \code{"3"}: 1:3 million.
}}

\item{moveCAN}{A logical \code{TRUE/FALSE} or a vector of coordinates
\code{c(lat, lon)}. It places the Canary Islands close to Spain's mainland.
Initial position can be adjusted using the vector of coordinates. See
\strong{Displacing the Canary Islands} in \code{\link[=esp_move_can]{esp_move_can()}}.}

\item{rawcols}{Logical. If \code{TRUE}, adds the raw columns of the resulting
object as provided by IGN.}
}
\value{
A \code{\link[sf:st_sf]{sf}} object.
}
\description{
Get
\href{https://en.wikipedia.org/wiki/Provinces_of_Spain}{provinces of Spain}
at a specified scale.
}
\details{
When using \code{prov} you can use and mix names and NUTS codes (levels 1, 2 or
3), ISO codes (corresponding to level 2 or 3) or \code{"cpro"} (see
\link{esp_codelist}).

Ceuta and Melilla are considered provinces in this dataset.

When calling a higher level (Autonomous Community or City, or NUTS 1), all
provinces of that level are added.
}
\examples{
\dontshow{if (esp_check_access()) withAutoprint(\{ # examplesIf}
library(ggplot2)

esp_get_ccaa_siane() |>
  dplyr::glimpse() |>
  ggplot() +
  geom_sf()
\dontshow{\}) # examplesIf}
}
\seealso{
Political and administrative boundary datasets:
\code{\link[=esp_get_capimun]{esp_get_capimun()}},
\code{\link[=esp_get_ccaa]{esp_get_ccaa()}},
\code{\link[=esp_get_ccaa_siane]{esp_get_ccaa_siane()}},
\code{\link[=esp_get_comarca]{esp_get_comarca()}},
\code{\link[=esp_get_countries_siane]{esp_get_countries_siane()}},
\code{\link{esp_get_gridmap}},
\code{\link[=esp_get_munic]{esp_get_munic()}},
\code{\link[=esp_get_munic_siane]{esp_get_munic_siane()}},
\code{\link[=esp_get_nuts]{esp_get_nuts()}},
\code{\link[=esp_get_prov]{esp_get_prov()}},
\code{\link{esp_get_simpl}},
\code{\link[=esp_get_spain]{esp_get_spain()}},
\code{\link[=esp_get_spain_siane]{esp_get_spain_siane()}},
\code{\link[=esp_siane_bulk_download]{esp_siane_bulk_download()}}

Datasets sourced from CartoBase ANE (Atlas Nacional de España):
\code{\link[=esp_get_capimun]{esp_get_capimun()}},
\code{\link[=esp_get_ccaa_siane]{esp_get_ccaa_siane()}},
\code{\link[=esp_get_countries_siane]{esp_get_countries_siane()}},
\code{\link[=esp_get_munic_siane]{esp_get_munic_siane()}},
\code{\link[=esp_get_spain_siane]{esp_get_spain_siane()}},
\code{\link[=esp_siane_bulk_download]{esp_siane_bulk_download()}}
}
\concept{political}
\concept{siane}
