% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esp-get-roads.R
\encoding{UTF-8}
\name{esp_get_roads}
\alias{esp_get_roads}
\title{Roads of Spain from SIANE}
\source{
CartoBase ANE (Atlas Nacional de España) provided by Instituto Geográfico Nacional (IGN),
\url{http://www.ign.es/web/ign/portal}. Years available are 2005 up to today.

Copyright:
\url{https://centrodedescargas.cnig.es/CentroDescargas/cartobase-ane}

Always acknowledge authorship using the following statements:
\enumerate{
\item When the original digital product is not modified or altered, use one of
the following statements:
\itemize{
\item CartoBase ANE 2006-2024 CC-BY 4.0 ign.es.
\item CartoBase ANE 2006-2024 CC-BY 4.0 Instituto Geográfico Nacional.
}
\item When a new product is generated:
\itemize{
\item Obra derivada de CartoBase ANE 2006-2024 CC-BY 4.0 ign.es.
}
}

Data distributed through the \code{sianedata} data branch, see
\url{https://github.com/rOpenSpain/mapSpain/tree/sianedata}.
}
\usage{
esp_get_roads(
  year = Sys.Date(),
  epsg = "4258",
  cache = TRUE,
  update_cache = FALSE,
  cache_dir = NULL,
  verbose = FALSE,
  moveCAN = TRUE
)
}
\arguments{
\item{year}{Ignored.}

\item{epsg}{Character string or number. Projection of the map: 4-digit
\href{https://epsg.io/}{EPSG code}. One of:
\itemize{
\item \code{"4258"}: \href{https://epsg.io/4258}{ETRS89}.
\item \code{"4326"}: \href{https://epsg.io/4326}{WGS84}.
\item \code{"3035"}: \href{https://epsg.io/3035}{ETRS89 / ETRS-LAEA}.
\item \code{"3857"}: \href{https://epsg.io/3857}{Pseudo-Mercator}.
}}

\item{cache}{Logical. Whether to cache downloaded files. Default is \code{TRUE}.
See \strong{Caching strategies} section in \code{\link[=esp_set_cache_dir]{esp_set_cache_dir()}}.}

\item{update_cache}{Logical. If \code{TRUE}, refreshes the cached file and forces
a new download. Defaults to \code{FALSE}.}

\item{cache_dir}{Character string. A path to a cache directory. See
\strong{Caching strategies} section in \code{\link[=esp_set_cache_dir]{esp_set_cache_dir()}}.}

\item{verbose}{A logical value. If \code{TRUE} displays informational messages.}

\item{moveCAN}{A logical \code{TRUE/FALSE} or a vector of coordinates
\code{c(lat, lon)}. It places the Canary Islands close to Spain's mainland.
Initial position can be adjusted using the vector of coordinates. See
\strong{Displacing the Canary Islands} in \code{\link[=esp_move_can]{esp_move_can()}}.}
}
\value{
A \code{\link[sf:st_sf]{sf}} object.
}
\description{
Object representing the main roads of Spain.
}
\examples{
\dontshow{if (esp_check_access()) withAutoprint(\{ # examplesIf}
\donttest{

country <- esp_get_spain()
roads <- esp_get_roads()

library(ggplot2)

ggplot(country) +
  geom_sf(fill = "grey90") +
  geom_sf(data = roads, aes(color = t_ctra_desc), show.legend = "line") +
  scale_color_manual(
    values = c("#003399", "#003399", "#ff0000", "#ffff00")
  ) +
  guides(color = guide_legend(direction = "vertical")) +
  theme_minimal() +
  labs(color = "Road type") +
  theme(legend.position = "bottom")
}
\dontshow{\}) # examplesIf}
}
\seealso{
Transport infrastructure datasets:
\code{\link[=esp_get_railway]{esp_get_railway()}}
}
\concept{infrastructure}
