% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esp-get-simpl.R
\encoding{UTF-8}
\name{esp_get_simpl}
\alias{esp_get_simpl}
\alias{esp_get_simpl_prov}
\alias{esp_get_simpl_ccaa}
\title{Simplified map of provinces and Autonomous Communities and Cities of Spain}
\source{
INE: PC-Axis files.
}
\usage{
esp_get_simpl_prov(
  prov = NULL,
  update_cache = FALSE,
  cache_dir = NULL,
  verbose = FALSE
)

esp_get_simpl_ccaa(
  ccaa = NULL,
  update_cache = FALSE,
  cache_dir = NULL,
  verbose = FALSE
)
}
\arguments{
\item{prov, ccaa}{Character. A vector of names, codes or both for provinces
and Autonomous Communities and Cities, or \code{NULL} to get all the data. See
\strong{Details}.}

\item{update_cache}{Logical. If \code{TRUE}, refreshes the cached file and forces
a new download. Defaults to \code{FALSE}.}

\item{cache_dir}{Character string. A path to a cache directory. See
\strong{Caching}.}

\item{verbose}{A logical value indicating whether to display informational
messages.}
}
\value{
A \code{\link[sf:st_sf]{sf}} object.
}
\description{
Simplified map with the boundaries of the provinces or Autonomous
Communities of Spain, as provided by the \strong{INE} (Instituto Nacional de
Estadística).
}
\details{
Results are provided \strong{without CRS}, as provided by the source.

You can use and mix names, ISO codes, \code{"codauto"} or \code{"cpro"} codes (see
\link{esp_codelist}) and NUTS codes of different levels.

When using a code corresponding to a higher level (for example,
\code{esp_get_prov("Andalucia")}), all the corresponding units of that level are
provided (in this case, all the provinces of Andalusia).
}
\section{Caching}{


Functions that download data store files in \code{cache_dir}. When \code{cache_dir}
is \code{NULL}, they use the active package cache, which defaults to a temporary
directory. Set \code{update_cache = TRUE} to replace an existing cached file.
See \strong{Caching strategies} in \code{\link[=esp_set_cache_dir]{esp_set_cache_dir()}} to configure a
persistent cache.
}

\examples{
\dontshow{if (esp_check_access()) withAutoprint(\{ # examplesIf}
\donttest{
prov_simp <- esp_get_simpl_prov()

library(ggplot2)

ggplot(prov_simp) +
  geom_sf(aes(fill = ine.ccaa.name)) +
  labs(fill = "Autonomous Communities and Cities")

# Provinces of a single Autonomous Community or City.

and_simple <- esp_get_simpl_prov("Andalucia")

ggplot(and_simple) +
  geom_sf()

# Autonomous Communities and Cities.

ccaa_simp <- esp_get_simpl_ccaa()

ggplot(ccaa_simp) +
  geom_sf() +
  geom_sf_text(aes(label = ine.ccaa.name), check_overlap = TRUE)
}
\dontshow{\}) # examplesIf}
}
\seealso{
Additional boundary datasets and representations:
\code{\link[=esp_get_comarca]{esp_get_comarca()}},
\code{\link{esp_get_gridmap}}
}
\concept{boundary_helpers}
\concept{political}
