% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esp-move-can.R
\encoding{UTF-8}
\name{esp_move_can}
\alias{esp_move_can}
\title{Displace an \code{sf} or \code{sfc} object located in the Canary Islands}
\usage{
esp_move_can(x, moveCAN = TRUE)
}
\arguments{
\item{x}{An \code{\link[sf:st_sf]{sf}} or \code{\link[sf:st_sfc]{sfc}} object.}

\item{moveCAN}{A logical \code{TRUE/FALSE} or a vector of coordinates
\code{c(lat, lon)}. It places the Canary Islands close to Spain's mainland.
Initial position can be adjusted using the vector of coordinates. See
\strong{Displacing the Canary Islands} in \code{\link[=esp_move_can]{esp_move_can()}}.}
}
\value{
An object of the same class and with the same CRS as \code{x}, displaced
accordingly.
}
\description{
Helper function to displace an external \code{sf} or \code{sfc} object,
potentially representing a location in the Canary Islands, to align it with
\CRANpkg{mapSpain} objects generated with \code{moveCAN = TRUE}.
}
\details{
This helper function makes it easier to represent objects located in the
Canary Islands that have been obtained from sources other than
\CRANpkg{mapSpain}.
}
\section{Displacing the Canary Islands}{
While \code{moveCAN} is useful for visualization, it will alter the actual
geographic position of the Canary Islands. When using the output for
spatial analysis or tiles (for example, with \code{\link[=esp_get_tiles]{esp_get_tiles()}} or
\code{\link[=addProviderEspTiles]{addProviderEspTiles()}}), set this option to \code{FALSE} to get the actual
coordinates instead of the modified ones.
}

\examples{
\dontshow{if (requireNamespace("ggplot2", quietly = TRUE)) withAutoprint(\{ # examplesIf}
library(sf)
teide <- data.frame(
  name = "Teide Peak",
  lon = -16.6437593,
  lat = 28.2722883
)

teide_sf <- st_as_sf(teide, coords = c("lon", "lat"), crs = 4326)

# A mapSpain object with moveCAN = TRUE is displaced.

esp <- esp_get_spain(moveCAN = c(13, 0))

library(ggplot2)

ggplot(esp) +
  geom_sf() +
  geom_sf(data = teide_sf, color = "red") +
  labs(
    title = "Canary Islands displaced",
    subtitle = "But not the external Teide object"
  )

# Displace the external object too.

teide_sf_disp <- esp_move_can(teide_sf, moveCAN = c(13, 0))

ggplot(esp) +
  geom_sf() +
  geom_sf(data = teide_sf_disp, color = "red") +
  labs(
    title = "Canary Islands displaced",
    subtitle = "And also the external Teide object"
  )
\dontshow{\}) # examplesIf}
}
\seealso{
Helpers for Canary Islands insets and displacement:
\code{\link[=esp_get_can_box]{esp_get_can_box()}}
}
\concept{can_helpers}
