% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer_spatraster.R
\name{layer_spatraster}
\alias{layer_spatraster}
\title{Deprecated: Use \code{tidyterra::geom_spatraster_rgb()}}
\usage{
layer_spatraster(data, ...)
}
\arguments{
\item{data}{Deprecated.}

\item{...}{Deprecated}
}
\value{
An error
}
\description{
This function is deprecated. Use \code{\link[tidyterra:geom_spatraster_rgb]{tidyterra::geom_spatraster_rgb()}} for
plot tiles with ggplot2.
}
\examples{
\dontshow{if (esp_check_access()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
\donttest{

# Get a SpatRaster

x <- esp_get_ccaa("Galicia")

tile <- esp_getTiles(x, "IDErioja")

class(tile)

library(ggplot2)
# Use tidyterra
library(tidyterra)

ggplot(x) +
  geom_spatraster_rgb(data = tile) +
  geom_sf(color = "yellow", fill = NA) +
  theme_minimal()
}
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[tidyterra:geom_spatraster_rgb]{tidyterra::geom_spatraster_rgb()}}.
}
\keyword{internal}
