% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapSpain-package.R
\docType{package}
\encoding{UTF-8}
\name{mapSpain-package}
\alias{mapSpain}
\alias{mapSpain-package}
\title{mapSpain: Administrative Boundaries and Static Map Tiles of Spain}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Administrative boundaries of Spain at several levels (Autonomous Communities and Cities, provinces, municipalities and 'NUTS'), based on 'GISCO' from 'Eurostat' \url{https://ec.europa.eu/eurostat/web/gisco} and 'CartoBase ANE' from 'Instituto Geográfico Nacional' \url{https://www.ign.es/}. Includes tools to download and process static map tiles and a 'leaflet' plugin for Spanish public administration tile providers.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ropenspain.github.io/mapSpain/}
  \item \url{https://github.com/rOpenSpain/mapSpain}
  \item Report bugs at \url{https://github.com/rOpenSpain/mapSpain/issues}
}

}
\author{
\strong{Maintainer}: Diego Hernangómez \email{diego.hernangomezherrero@gmail.com} (\href{https://orcid.org/0000-0001-8457-4658}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Diego Hernangómez \email{diego.hernangomezherrero@gmail.com} (\href{https://orcid.org/0000-0001-8457-4658}{ORCID}) [copyright holder]
}

Other contributors:
\itemize{
  \item Eurostat (\href{https://ror.org/033d3q980}{ROR}) [copyright holder]
  \item Instituto Geográfico Nacional (\href{https://ror.org/03yycdv57}{ROR}) [copyright holder]
  \item Francisco J. Goerlich (\href{https://orcid.org/0000-0003-1626-525X}{ORCID}) [contributor]
}

}
\keyword{internal}
