% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapSpain-package.R
\docType{package}
\name{mapSpain-package}
\alias{mapSpain}
\alias{mapSpain-package}
\title{mapSpain: Administrative Boundaries of Spain}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Administrative Boundaries of Spain at several levels (Autonomous Communities, Provinces, Municipalities) based on the 'GISCO' 'Eurostat' database \url{https://ec.europa.eu/eurostat/web/gisco} and 'CartoBase SIANE' from 'Instituto Geografico Nacional' \url{https://www.ign.es/}. It also provides a 'leaflet' plugin and the ability of downloading and processing static tiles.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ropenspain.github.io/mapSpain/}
  \item \url{https://github.com/rOpenSpain/mapSpain}
  \item Report bugs at \url{https://github.com/rOpenSpain/mapSpain/issues}
}

}
\author{
\strong{Maintainer}: Diego Hernangómez \email{diego.hernangomezherrero@gmail.com} (\href{https://orcid.org/0000-0001-8457-4658}{ORCID}) [copyright holder]

Other contributors:
\itemize{
  \item Timothée Giraud \email{timothee.giraud@cnrs.fr} (\href{https://orcid.org/0000-0002-1932-3323}{ORCID}) (Adaptation of slippymath functions) [contributor]
  \item Miles McBain (\href{https://orcid.org/0000-0003-2865-2548}{ORCID}) (Slippy map tiles functions, from slippymath package) [copyright holder]
  \item Eurostat (\href{https://ror.org/033d3q980}{ROR}) [copyright holder]
  \item Instituto Geográfico Nacional (\href{https://ror.org/03yycdv57}{ROR}) [copyright holder]
}

}
\keyword{internal}
