% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapSpain-package.R
\docType{package}
\name{mapSpain-package}
\alias{mapSpain-package}
\alias{mapSpain}
\title{mapSpain package}
\source{
\href{https://ec.europa.eu/eurostat/web/gisco/geodata/reference-data/administrative-units-statistical-units}{GISCO webpage}
}
\description{
\if{html}{\figure{logo.png}{options: width=120 alt="mapSpain logo" align='right'}}
This package provides Administrative Boundaries of Spain based on
the GISCO (Geographic Information System of the Commission) Eurostat
database.
}
\details{
\tabular{ll}{
Package: \tab mapSpain \cr
Type: \tab Package\cr
Version: \tab See sessionInfo() or DESCRIPTION file\cr
Date: \tab 2020\cr
License: \tab GPL-3\cr
LazyLoad: \tab yes\cr
}
}
\note{
COPYRIGHT NOTICE

When data downloaded from
\href{https://ec.europa.eu/eurostat/web/gisco/geodata/reference-data/administrative-units-statistical-units}{this page}
is used in any printed or electronic publication,
in addition to any other provisions
applicable to the whole Eurostat website,
data source will have to be acknowledged
in the legend of the map and
in the introductory page of the publication
with the following copyright notice:
\itemize{
	\item EN: (C) EuroGeographics for the administrative boundaries
	\item FR: (C) EuroGeographics pour les limites administratives
	\item DE: (C) EuroGeographics bezüglich der Verwaltungsgrenzen
}
For publications in languages other than
English, French or German,
the translation of the copyright notice
in the language of the publication shall be used.

If you intend to use the data commercially,
please contact EuroGeographics for
information regarding their license agreements.
}
\references{
See \code{citation("mapSpain")}
}
\seealso{
Useful links:
 \itemize{
 \item \url{https://dieghernan.github.io/mapSpain/}
 \item \url{https://github.com/dieghernan/mapSpain}
 \item Report bugs at \url{https://github.com/dieghernan/mapSpain/issues}
 }
}
\author{
dieghernan, \url{https://github.com/dieghernan/}
}
\keyword{package}
