% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pobmun19}
\alias{pobmun19}
\title{Population by municipality (2019)}
\source{
INE: Instituto Nacional de Estadistica \url{https://www.ine.es/}
}
\description{
A data frame with 8,131
rows containing the population data by municipality in Spain (2019).
}
\examples{
data("pobmun19")

library(tibble)

as_tibble(pobmun19)
}
\concept{datasets}
