% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pobmun19}
\alias{pobmun19}
\title{Database with the population of Spain by municipality (2019)}
\format{
An example \code{data.frame} object with
8,131 rows containing the
population data by municipality in Spain (2019).

\describe{
\item{cpro}{INE code of the province.}
\item{provincia}{name of the province.}
\item{cmun}{INE code of the municipality.}
\item{name}{Name of the municipality.}
\item{pob19}{Overall population (2019)}
\item{men}{Men population (2019)}
\item{women}{Women population (2019)}
}
}
\source{
INE: Instituto Nacional de Estadistica \url{https://www.ine.es/}
}
\description{
Database with the population of Spain by municipality (2019)
}
\examples{
data("pobmun19")
}
\seealso{
Other datasets: 
\code{\link{esp_codelist}},
\code{\link{esp_munic.sf}},
\code{\link{esp_nuts.sf}},
\code{\link{esp_nuts_2024}},
\code{\link{esp_tiles_providers}}
}
\concept{datasets}
